# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ユーザー管理サービスAPIの機能",
        "title": "ユーザー管理サービスAPI",
        "contact": {},
        "version": "v0.1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/v1/health": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "healthcheck"
                ],
                "summary": "死活監視用",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/system.Response"
                        }
                    }
                }
            }
        },
        "/v1/users": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetUsers"
                ],
                "summary": "ユーザー一覧を取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.ResponseUser"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EditUser"
                ],
                "summary": "ユーザー情報を編集",
                "parameters": [
                    {
                        "description": "ユーザー情報",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/user.RequestUserParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.Response"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DeleteUser"
                ],
                "summary": "ユーザー情報を削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "request",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.Response"
                        }
                    }
                }
            }
        },
        "/v1/users/:id": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetUserById"
                ],
                "summary": "ユーザーの詳細情報を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "request",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseUser"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "system.Response": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "user.RequestUserParam": {
            "type": "object",
            "properties": {
                "birth_day": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "extra": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "post_office_number": {
                    "type": "string"
                },
                "pref": {
                    "type": "string"
                },
                "user___id": {
                    "type": "string"
                }
            }
        },
        "user.Response": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "user.ResponseUser": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/user.ResponseUserModel"
                }
            }
        },
        "user.ResponseUserModel": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birth_day": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "post_office_number": {
                    "type": "string"
                },
                "user___id": {
                    "type": "string"
                }
            }
        }
    }
}